Remove onnxruntime extensions from list of gitmodules (#17615)
The extensions submodule was removed in [this
PR](https://github.com/microsoft/onnxruntime/pull/17097) but not deleted
from the list of git modules. This causes breaks in code ingesting ORT
that references the git modules for an accurate list of submodules.

This change removes the extensions from the list of git modules to
resolve this issue.